- In the NPCSpawner, try multiple (50) times to find a valid spawn position that doesn't contain any walls instead of hoping the entity updating takes care of that (should improve NPC spawning positions) - Clean up usages of Rectangle, especially by letting the constructors handle casting to int (amazing how many times I did that before I thought to add a new constructor overload) and by using .Inflate() where appropriate - Fix buggy looking code for determining max ranged attack distance

+++ netgore/trunk/DemoGame.Client/Controls/SkillCastProgressBar.cs
@@ -48,7 +48,7 @@ namespace DemoGame.Client
             var percent = elapsedTime / _currentCastTime;
             var startDrawPos = ScreenPosition + new Vector2(Border.LeftWidth, Border.TopHeight);
             var drawWidth = ClientSize.X * Math.Min(percent, 1);
-            var r = new Rectangle((int)startDrawPos.X, (int)startDrawPos.Y, (int)drawWidth, (int)ClientSize.Y);
+            var r = new Rectangle(startDrawPos.X, startDrawPos.Y, drawWidth, ClientSize.Y);
             RenderRectangle.Draw(spriteBatch, r, new Color(255, 0, 0, 200));
 
             // Draw the name
