+++ netgore/trunk/NetGore.Tests/NetGore/SpatialAggregateTests.cs
@@ -52,7 +52,7 @@ namespace NetGore.Tests.NetGore
             Entity someEntity;
             foreach (var spatial in GetSpatials(out someEntity))
             {
-                var all = spatial.GetMany(new Rectangle(0, 0, (int)SpatialSize.X, (int)SpatialSize.Y));
+                var all = spatial.GetMany(new Rectangle(0, 0, SpatialSize.X, SpatialSize.Y));
                 foreach (var a in all)
                 {
                     Assert.IsTrue(spatial.CollectionContains(a));
@@ -76,7 +76,7 @@ namespace NetGore.Tests.NetGore
             Entity someEntity;
             foreach (var spatial in GetSpatials(out someEntity))
             {
-                var all = spatial.GetMany(new Rectangle(0, 0, (int)SpatialSize.X, (int)SpatialSize.Y));
+                var all = spatial.GetMany(new Rectangle(0, 0, SpatialSize.X, SpatialSize.Y));
                 var distinct = all.Distinct();
                 Assert.AreEqual(distinct.Count(), all.Count());
             }
